Navigate cross-feature via concrete keys from -navigation modules
Replace the field-setting navigation dance (presenter sets a UiState flag
-> UI LaunchedEffect -> injected lambda -> central when mapping intent to
key) with presenters that add concrete nav keys to the Backstack directly.

Introduce feature-x-navigation sister modules hosting only the keys meant
to be reached cross-feature, which the build plugin exempts from the
feature->feature dependency ban. Rewrite HelpCenterPresenter to map every
quick-link destination straight to a key, dropping destinationToNavigate
and the ClearNavigation event, and remove the central when lambda in the
app's HedvigEntryProvider.
