+++ app/app/build.gradle.kts
@@ -189,13 +189,16 @@ dependencies {
   implementation(projects.featureChat)
   implementation(projects.featureChipId)
   implementation(projects.featureChooseTier)
+  implementation(projects.featureChooseTierNavigation)
   implementation(projects.featureClaimChat)
   implementation(projects.featureClaimDetails)
   implementation(projects.featureClaimHistory)
   implementation(projects.featureConnectPaymentTrustly)
+  implementation(projects.featureConnectPaymentTrustlyNavigation)
   implementation(projects.featureCrossSellSheet)
   implementation(projects.featureDeleteAccount)
   implementation(projects.featureEditCoinsured)
+  implementation(projects.featureEditCoinsuredNavigation)
   implementation(projects.featureFlags)
   implementation(projects.featureForever)
   implementation(projects.featureHelpCenter)
@@ -205,13 +208,16 @@ dependencies {
   implementation(projects.featureInsurances)
   implementation(projects.featureLogin)
   implementation(projects.featureMovingflow)
+  implementation(projects.featureMovingflowNavigation)
 
   implementation(projects.featureRemoveAddons)
   implementation(projects.featurePayoutAccount)
   implementation(projects.featurePayments)
   implementation(projects.featureProfile)
   implementation(projects.featureTerminateInsurance)
+  implementation(projects.featureTerminateInsuranceNavigation)
   implementation(projects.featureTravelCertificate)
+  implementation(projects.featureTravelCertificateNavigation)
   implementation(projects.foreverUi)
   implementation(projects.initializable)
   implementation(projects.languageCore)
