+++ app/feature/feature-choose-tier-navigation/src/commonMain/kotlin/com/hedvig/android/feature/change/tier/navigation/StartTierFlowChooseInsuranceKey.kt
@@ -1,0 +1,7 @@
+package com.hedvig.android.feature.change.tier.navigation
+
+import com.hedvig.android.navigation.common.HedvigNavKey
+import kotlinx.serialization.Serializable
+
+@Serializable
+data object StartTierFlowChooseInsuranceKey : HedvigNavKey
