+++ app/feature/feature-connect-payment-trustly-navigation/src/commonMain/kotlin/com/hedvig/android/feature/connect/payment/trustly/ui/TrustlyKey.kt
@@ -1,0 +1,7 @@
+package com.hedvig.android.feature.connect.payment.trustly.ui
+
+import com.hedvig.android.navigation.common.HedvigNavKey
+import kotlinx.serialization.Serializable
+
+@Serializable
+data object TrustlyKey : HedvigNavKey
