+++ app/feature/feature-connect-payment-trustly/build.gradle.kts
@@ -39,6 +39,7 @@ dependencies {
   implementation(projects.coreCommonPublic)
   implementation(projects.coreResources)
   implementation(projects.designSystemHedvig)
+  implementation(projects.featureConnectPaymentTrustlyNavigation)
   implementation(projects.moleculePublic)
   implementation(projects.navigationCommon)
   implementation(projects.navigationCompose)
