+++ app/feature/feature-edit-coinsured/build.gradle.kts
@@ -37,6 +37,7 @@ dependencies {
   implementation(projects.coreUiData)
   implementation(projects.designSystemApi)
   implementation(projects.designSystemHedvig)
+  implementation(projects.featureEditCoinsuredNavigation)
   implementation(projects.moleculePublic)
   implementation(projects.navigationCommon)
   implementation(projects.navigationCompose)
