+++ app/feature/feature-edit-coinsured/src/main/kotlin/com/hedvig/android/feature/editcoinsured/navigation/EditCoInsuredDestinations.kt
@@ -10,25 +10,6 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class CoInsuredAddInfoKey(
-  val contractId: String,
-  val type: CoInsuredFlowType,
-) : HedvigNavKey
-
-@Serializable
-data class CoInsuredAddOrRemoveKey(
-  val contractId: String,
-  val type: CoInsuredFlowType,
-) : HedvigNavKey
-
-@Serializable
-data class EditCoInsuredTriageKey(
-  @SerialName("contractId")
-  val contractId: String? = null,
-  val type: CoInsuredFlowType = CoInsuredFlowType.CoInsured,
-) : HedvigNavKey
-
-@Serializable
 internal data class EditCoOwnersTriageDeepLinkKey(
   @SerialName("contractId")
   val contractId: String? = null,
