+++ app/feature/feature-help-center/src/commonMain/kotlin/com/hedvig/android/feature/help/center/HelpCenterPresenter.kt
@@ -11,8 +11,13 @@ import androidx.compose.runtime.setValue
 import arrow.core.NonEmptyList
 import arrow.core.merge
 import arrow.core.toNonEmptyListOrNull
+import com.hedvig.android.data.coinsured.CoInsuredFlowType
 import com.hedvig.android.data.conversations.HasAnyActiveConversationUseCase
-import com.hedvig.android.feature.help.center.HelpCenterEvent.ClearNavigation
+import com.hedvig.android.feature.change.tier.navigation.StartTierFlowChooseInsuranceKey
+import com.hedvig.android.feature.connect.payment.trustly.ui.TrustlyKey
+import com.hedvig.android.feature.editcoinsured.navigation.CoInsuredAddInfoKey
+import com.hedvig.android.feature.editcoinsured.navigation.CoInsuredAddOrRemoveKey
+import com.hedvig.android.feature.editcoinsured.navigation.EditCoInsuredTriageKey
 import com.hedvig.android.feature.help.center.HelpCenterEvent.ClearSearchQuery
 import com.hedvig.android.feature.help.center.HelpCenterEvent.NavigateToQuickAction
 import com.hedvig.android.feature.help.center.HelpCenterEvent.OnDismissQuickActionDialog
@@ -28,11 +33,26 @@ import com.hedvig.android.feature.help.center.data.GetPuppyGuideUseCase
 import com.hedvig.android.feature.help.center.data.GetQuickLinksUseCase
 import com.hedvig.android.feature.help.center.data.InnerHelpCenterDestination
 import com.hedvig.android.feature.help.center.data.QuickLinkDestination
+import com.hedvig.android.feature.help.center.data.QuickLinkDestination.OuterDestination.ChooseInsuranceForEditCoInsured
+import com.hedvig.android.feature.help.center.data.QuickLinkDestination.OuterDestination.ChooseInsuranceForEditCoOwners
+import com.hedvig.android.feature.help.center.data.QuickLinkDestination.OuterDestination.QuickLinkChangeAddress
+import com.hedvig.android.feature.help.center.data.QuickLinkDestination.OuterDestination.QuickLinkChangeTier
+import com.hedvig.android.feature.help.center.data.QuickLinkDestination.OuterDestination.QuickLinkCoInsuredAddInfo
+import com.hedvig.android.feature.help.center.data.QuickLinkDestination.OuterDestination.QuickLinkCoInsuredAddOrRemove
+import com.hedvig.android.feature.help.center.data.QuickLinkDestination.OuterDestination.QuickLinkCoOwnerAddInfo
+import com.hedvig.android.feature.help.center.data.QuickLinkDestination.OuterDestination.QuickLinkCoOwnerAddOrRemove
+import com.hedvig.android.feature.help.center.data.QuickLinkDestination.OuterDestination.QuickLinkConnectPayment
+import com.hedvig.android.feature.help.center.data.QuickLinkDestination.OuterDestination.QuickLinkTermination
+import com.hedvig.android.feature.help.center.data.QuickLinkDestination.OuterDestination.QuickLinkTravelCertificate
 import com.hedvig.android.feature.help.center.model.QuickAction
 import com.hedvig.android.feature.help.center.navigation.EmergencyKey
 import com.hedvig.android.feature.help.center.navigation.FirstVetKey
+import com.hedvig.android.feature.movingflow.SelectContractForMovingKey
+import com.hedvig.android.feature.terminateinsurance.navigation.TerminateInsuranceKey
+import com.hedvig.android.feature.travelcertificate.navigation.TravelCertificateKey
 import com.hedvig.android.molecule.public.MoleculePresenter
 import com.hedvig.android.molecule.public.MoleculePresenterScope
+import com.hedvig.android.navigation.common.HedvigNavKey
 import com.hedvig.android.navigation.compose.Backstack
 import com.hedvig.android.navigation.compose.add
 import kotlinx.coroutines.flow.collect
@@ -54,8 +74,6 @@ internal sealed interface HelpCenterEvent {
 
   data class NavigateToQuickAction(val destination: QuickLinkDestination) : HelpCenterEvent
 
-  data object ClearNavigation : HelpCenterEvent
-
   data object ReloadFAQAndQuickLinks : HelpCenterEvent
 }
 
@@ -66,7 +84,6 @@ internal data class HelpCenterUiState(
   val selectedQuickAction: QuickAction?,
   val search: Search?,
   val showNavigateToInboxButton: Boolean,
-  val destinationToNavigate: QuickLinkDestination.OuterDestination? = null,
   val puppyGuide: PuppyGuidePresentation?,
 ) {
   data class QuickLink(val quickAction: QuickAction)
@@ -153,26 +170,62 @@ internal class HelpCenterPresenter(
           }
         }
 
-        ClearNavigation -> {
-          selectedQuickAction = null
-          currentState = currentState.copy(destinationToNavigate = null)
-        }
-
         is NavigateToQuickAction -> {
           selectedQuickAction = null
-          when (val destination = event.destination) {
+          val key: HedvigNavKey = when (val destination = event.destination) {
             is InnerHelpCenterDestination.FirstVet -> {
-              backstack.add(FirstVetKey(destination.sections))
+              FirstVetKey(destination.sections)
             }
 
             is InnerHelpCenterDestination.QuickLinkSickAbroad -> {
-              backstack.add(EmergencyKey(destination.deflectData))
+              EmergencyKey(destination.deflectData)
             }
 
-            is QuickLinkDestination.OuterDestination -> {
-              currentState = currentState.copy(destinationToNavigate = destination)
+            QuickLinkChangeAddress -> {
+              SelectContractForMovingKey
+            }
+
+            is QuickLinkCoInsuredAddInfo -> {
+              CoInsuredAddInfoKey(destination.contractId, CoInsuredFlowType.CoInsured)
+            }
+
+            is QuickLinkCoInsuredAddOrRemove -> {
+              CoInsuredAddOrRemoveKey(destination.contractId, CoInsuredFlowType.CoInsured)
+            }
+
+            is QuickLinkCoOwnerAddInfo -> {
+              CoInsuredAddInfoKey(destination.contractId, CoInsuredFlowType.CoOwners)
+            }
+
+            is QuickLinkCoOwnerAddOrRemove -> {
+              CoInsuredAddOrRemoveKey(destination.contractId, CoInsuredFlowType.CoOwners)
+            }
+
+            QuickLinkConnectPayment -> {
+              TrustlyKey
+            }
+
+            QuickLinkTermination -> {
+              TerminateInsuranceKey(null)
+            }
+
+            QuickLinkTravelCertificate -> {
+              TravelCertificateKey
+            }
+
+            QuickLinkChangeTier -> {
+              StartTierFlowChooseInsuranceKey
+            }
+
+            ChooseInsuranceForEditCoInsured -> {
+              EditCoInsuredTriageKey()
+            }
+
+            ChooseInsuranceForEditCoOwners -> {
+              EditCoInsuredTriageKey(type = CoInsuredFlowType.CoOwners)
             }
           }
+          backstack.add(key)
         }
 
         HelpCenterEvent.ReloadFAQAndQuickLinks -> {
