+++ app/feature/feature-movingflow-navigation/src/commonMain/kotlin/com/hedvig/android/feature/movingflow/SelectContractForMovingKey.kt
@@ -1,0 +1,7 @@
+package com.hedvig.android.feature.movingflow
+
+import com.hedvig.android.navigation.common.HedvigNavKey
+import kotlinx.serialization.Serializable
+
+@Serializable
+data object SelectContractForMovingKey : HedvigNavKey
