+++ app/feature/feature-movingflow/build.gradle.kts
@@ -34,6 +34,7 @@ dependencies {
   implementation(projects.dataProductVariantPublic)
   implementation(projects.designSystemHedvig)
   implementation(projects.featureFlags)
+  implementation(projects.featureMovingflowNavigation)
   implementation(projects.moleculePublic)
   implementation(projects.navigationCommon)
   implementation(projects.navigationCompose)
