+++ app/feature/feature-movingflow/src/main/kotlin/com/hedvig/android/feature/movingflow/MovingFlowEntries.kt
@@ -27,9 +27,6 @@ import kotlinx.datetime.LocalDate
 import kotlinx.serialization.Serializable
 
 @Serializable
-data object SelectContractForMovingKey : HedvigNavKey
-
-@Serializable
 internal data class HousingTypeKey(val moveIntentId: String) : HedvigNavKey
 
 @Serializable
