+++ app/feature/feature-terminate-insurance/build.gradle.kts
@@ -37,6 +37,7 @@ dependencies {
   implementation(projects.dataTermination)
   implementation(projects.designSystemHedvig)
   implementation(projects.featureFlags)
+  implementation(projects.featureTerminateInsuranceNavigation)
   implementation(projects.languageCore)
   implementation(projects.moleculePublic)
   implementation(projects.navigationCommon)
