+++ app/feature/feature-travel-certificate-navigation/build.gradle.kts
@@ -1,0 +1,20 @@
+plugins {
+  id("hedvig.multiplatform.library")
+  id("hedvig.multiplatform.library.android")
+  id("hedvig.gradle.plugin")
+}
+
+hedvig {
+  serialization()
+  navKeys()
+}
+
+kotlin {
+  sourceSets {
+    commonMain.dependencies {
+      implementation(libs.kotlinx.serialization.core)
+      implementation(projects.coreCommonPublic)
+      implementation(projects.navigationCommon)
+    }
+  }
+}
