+++ app/feature/feature-travel-certificate-navigation/src/commonMain/kotlin/com/hedvig/android/feature/travelcertificate/navigation/TravelCertificateKey.kt
@@ -1,0 +1,8 @@
+package com.hedvig.android.feature.travelcertificate.navigation
+
+import com.hedvig.android.navigation.common.CrossSellEligibleDestination
+import com.hedvig.android.navigation.common.HedvigNavKey
+import kotlinx.serialization.Serializable
+
+@Serializable
+data object TravelCertificateKey : HedvigNavKey, CrossSellEligibleDestination
