+++ app/feature/feature-travel-certificate/build.gradle.kts
@@ -33,6 +33,7 @@ dependencies {
   implementation(projects.designSystemHedvig)
   implementation(projects.featureFlags)
   implementation(projects.languageCore)
+  implementation(projects.featureTravelCertificateNavigation)
   implementation(projects.moleculePublic)
   implementation(projects.navigationActivity)
   implementation(projects.navigationCommon)
